pyproject.toml
[tool.poetry]
name = "PawpularityPipeline"
version = "0.1.0"
description = ""
authors = ["Nikolay Lyudkevich <nikolai.lyudkevich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
kaggle = "^1.5.12"
numpy = "1.19.3"
pandas = "1.3.2"
PyYAML = "^6.0"
tensorflow = "2.5.0"
mlflow = "^1.22.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
